Update cron_date_input() method:
Fixed a typo in the output string, and turned off the abreviation
feature of the choice_input being used to collect data. Also I had to
fix an issue that when doing either MONTHLY or WEEKLY frequencies it
would crash due to trying to initialize the CronDate with a variable
that had not been declared yet.