Fixed an issue with version checking:
Previously the installer would have issues and fail if there was no MC
server installed when the install command would run. This was due to the
update type being MAJOR, and having the allow_major_updates config set
to False. This has been fixed. So when you first run the program,
without any version passed in (or installed) it will install properly
rather than error out.